got my classifier to work
the problem was with using np.zeros, and not truncated the data array
of the validation features. It all works clean now, and I got it to
print out the required results file. It didn't do any better than the
cross validation said it would, but I guess that's too be expected.